+++ src/wallet/api/CMakeLists.txt
@@ -30,27 +30,42 @@
 
 set(CMAKE_ARCHIVE_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/lib)
 
-set(wallet_sources
-  wallet2.cpp
-  wallet_args.cpp
-  node_rpc_proxy.cpp)
+set(wallet_api_sources
+  wallet.cpp
+  wallet_manager.cpp
+  transaction_info.cpp
+  transaction_history.cpp
+  pending_transaction.cpp
+  utils.cpp
+  address_book.cpp
+  subaddress.cpp
+  subaddress_account.cpp
+  unsigned_transaction.cpp)
 
-set(wallet_private_headers
-  wallet2.h
-  wallet_args.h
-  wallet_errors.h
-  wallet_rpc_server.h
-  wallet_rpc_server_commands_defs.h
-  wallet_rpc_server_error_codes.h
-  node_rpc_proxy.h)
+set(wallet_api_headers
+    wallet2_api.h)
 
-monero_private_headers(wallet
-  ${wallet_private_headers})
-monero_add_library(wallet
-  ${wallet_sources}
-  ${wallet_private_headers})
-target_link_libraries(wallet
+set(wallet_api_private_headers
+  wallet.h
+  wallet_manager.h
+  transaction_info.h
+  transaction_history.h
+  pending_transaction.h
+  common_defines.h
+  address_book.h
+  subaddress.h
+  subaddress_account.h
+  unsigned_transaction.h)
+
+monero_private_headers(wallet_api
+  ${wallet_api_private_headers})
+monero_add_library(wallet_api
+  ${wallet_api_sources}
+  ${wallet_api_headers}
+  ${wallet_api_private_headers})
+target_link_libraries(wallet_api
   PUBLIC
+    wallet
     common
     cryptonote_core
     mnemonics
@@ -63,67 +78,12 @@ target_link_libraries(wallet
   PRIVATE
     ${EXTRA_LIBRARIES})
 
-set(wallet_rpc_sources
-  wallet_rpc_server.cpp)
 
-set(wallet_rpc_headers)
-
-set(wallet_rpc_private_headers
-  wallet_rpc_server.h)
-
-monero_private_headers(wallet_rpc_server
-  ${wallet_rpc_private_headers})
-monero_add_executable(wallet_rpc_server
-  ${wallet_rpc_sources}
-  ${wallet_rpc_headers}
-  ${wallet_rpc_private_headers})
-
-target_link_libraries(wallet_rpc_server
-  PRIVATE
-    wallet
-    epee
-    rpc
-    cryptonote_core
-    cncrypto
-    common
-    version
-    ${Boost_CHRONO_LIBRARY}
-    ${Boost_PROGRAM_OPTIONS_LIBRARY}
-    ${Boost_FILESYSTEM_LIBRARY}
-    ${Boost_THREAD_LIBRARY}
-    ${CMAKE_THREAD_LIBS_INIT}
-    ${EXTRA_LIBRARIES})
-set_property(TARGET wallet_rpc_server
-  PROPERTY
-    OUTPUT_NAME "monero-wallet-rpc")
-install(TARGETS wallet_rpc_server DESTINATION bin)
-
-
-# build and install libwallet_merged only if we building for GUI
-if (BUILD_GUI_DEPS)
-    set(libs_to_merge
-            wallet_api
-            wallet
-            cryptonote_core
-            cryptonote_basic
-            mnemonics
-            common
-            cncrypto
-            ringct
-            checkpoints
-            version)
-
-    foreach(lib ${libs_to_merge})
-        list(APPEND objlibs $<TARGET_OBJECTS:obj_${lib}>) # matches naming convention in src/CMakeLists.txt
-    endforeach()
-    add_library(wallet_merged STATIC ${objlibs})
-    if(IOS)
-        set(lib_folder lib-${ARCH})
-    else()
-        set(lib_folder lib)
-    endif()
-    install(TARGETS wallet_merged
-        ARCHIVE DESTINATION ${lib_folder})
-
-    add_subdirectory(api)
+if(IOS)
+    set(lib_folder lib-${ARCH})
+else()
+    set(lib_folder lib)
 endif()
+
+install(FILES ${wallet_api_headers}
+    DESTINATION include/wallet)
