+++ src/wallet/api/transaction_history.h
@@ -28,7 +28,7 @@
 //
 // Parts of this file are originally copyright (c) 2012-2013 The Cryptonote developers
 
-#include "wallet/wallet2_api.h"
+#include "wallet/api/wallet2_api.h"
 #include <boost/thread/shared_mutex.hpp>
 
 namespace Monero {
