+++ src/wallet/wallet2.cpp
@@ -39,7 +39,6 @@ using namespace epee;
 
 #include "cryptonote_config.h"
 #include "wallet2.h"
-#include "wallet2_api.h"
 #include "cryptonote_basic/cryptonote_format_utils.h"
 #include "rpc/core_rpc_server_commands_defs.h"
 #include "misc_language.h"
