+++ tests/libwallet_api_tests/CMakeLists.txt
@@ -39,6 +39,7 @@ add_executable(libwallet_api_tests
 
 target_link_libraries(libwallet_api_tests
   PRIVATE
+    wallet_api
     wallet
     version
     epee
